Dropped 'pgperf'
pgperf was the customized verion of GNU gperf that came with LambdaMOO.
GNU gperf has supported case-insensitive lookup tables since 2002, and
pgperf didn't compile anymore anyways, due to its reliance on varargs.h
which is no longer supported.

(transferred comment from Makefile)

Pavel Curtis added a new command line option, '-l', to make the
generated hash table case-insensitive.

In 2002, the option '--ignore-case' was added to GNU gperf, and by
2011 the 'pgperf' shipped with LambdaMOO no longer compiled due to
obsolesence (it used varargs.h, which is no longer supported).

Also, the '-a' command line flag was replaced by '--language'
and the flag '-p' was dropped (and no explanation given in gperf's
ChangeLog).  I've expanded the short options into long options here
for readability.  For the sake of history, this was the original
incantation:

  pgperf -aCIptT -k1,3,$$ keywords.gperf

--sw July 2011

[Note: this change was modified to work with C instead -- Kenny Root]