Detect symbols defined in headers more directly.
On El Capitan and other Darwin-based distros, some functions defined in system
headers are declared in ways more complicated than the regex in `configure` was
prepared for. Rather than try to make the regex more complex, `configure` now
directly attempts to compile and link code that calls the function under test.
This is a more reliable way to detect whether the function is declared with a
suitable signature.